refactor out main program loop into its own method
startYappin now consists of print statements and a loop, but no switch case statement
handleInstruction consists of a big switch case statement
