+++ src/main/java/yapper/Yapper.java
@@ -1,13 +1,33 @@
 package yapper;
 
+import java.util.Scanner;
+import yapper.stringClasses.StringStorage;
+
 // Main function for Yapper
 public class Yapper {
     public static final int maxCapacity = 100;
+
+    // Main ChatBot Loop
+    public static void startYappin(TaskManager taskManager) {
+        System.out.println(StringStorage.START_UP_MESSAGE);
+        System.out.println(StringStorage.HELP_MESSAGE); // TODO
+
+        Scanner scanner = new Scanner(System.in);
+        while (true) {
+            String userInputString = scanner.nextLine();
+            if (userInputString.trim().equals("bye")) {
+                break;
+            }
+            UserInterface.handleInstruction(taskManager, userInputString);
+        }
+
+        System.out.println(StringStorage.SHUT_DOWN_MESSAGE);
+    }
 
     public static void main(String[] args) {
         // Initialize
         TaskManager taskManager = new TaskManager(maxCapacity);
         // Startup ChatBot Program
-        UserInterface.startYappin(taskManager);
+        startYappin(taskManager);
     }
 }
